requirements.txt
annotated-types
anyio
appnope
asttokens
backoff
boto3
botocore
certifi
charset-normalizer
comm
debugpy
decorator
distro
executing
h11
httpcore
httpx
idna
ipykernel
ipython
ipython-pygments-lexers
jedi
jiter
jmespath
jupyter-client
jupyter-core
levenshtein
matplotlib-inline
nest-asyncio
numpy
openai
packaging
pandas
parso
pexpect
platformdirs
pre-commit
prompt-toolkit
psutil
ptyprocess
pure-eval
pydantic
pydantic-core
pygments
python-dateutil
python-dotenv
python-levenshtein
pytz
pyzmq
rapidfuzz
requests
s3transfer
six
sniffio
stack-data
tornado
tqdm
traitlets
typing-extensions
tzdata
urllib3
wcwidth
mistralai
pillow